Project/requirements.txt
amqp==5.0.3
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.3
django-enumfield==2.0.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
huey==2.3.0
idna==2.10
importlib-metadata==3.4.0
kombu==5.0.2
paho-mqtt==1.5.1
prompt-toolkit==3.0.11
PyJWT==1.7.1
pytz==2020.4
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
